requirements.txt
Django==5.2
djangorestframework==3.16.0
drf-yasg==1.21.10
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
pillow==11.2.1
psycopg2-binary==2.9.10